Screens refactor finished.
