+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/data/repository/Constants.kt
@@ -1,5 +1,5 @@
 package com.pinheiro.marvelhqs.data.repository
 
 object Constants {
-    const val NETWORK_BASE_URL = "https://gateway.marvel.com/"
+    const val NETWORK_BASE_URL = "https://gateway.marvel.com/" // colocar como variável de ambiente
 }
