+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/data/repository/db/realm/model/ComicRealm.kt
@@ -11,4 +11,5 @@ class ComicRealm : RealmObject {
     var variantDescription: String? = null
     var description: String? = null
     var pageCount: Int? = null
+    var thumbnail: String? = null
 }
