+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presentation/MainApplication.kt
@@ -1,0 +1,21 @@
+package com.pinheiro.marvelhqs.presentation
+
+import android.app.Application
+import com.pinheiro.marvelhqs.di.appModule
+import org.koin.android.ext.koin.androidContext
+import org.koin.android.ext.koin.androidLogger
+import org.koin.core.context.GlobalContext.startKoin
+
+class MainApplication : Application() {
+
+
+    override fun onCreate() {
+        super.onCreate()
+        startKoin{
+            androidLogger()
+            androidContext(this@MainApplication)
+            modules(appModule)
+        }
+
+    }
+}
