+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presentation/interfaces/Paginator.kt
@@ -1,0 +1,6 @@
+package com.pinheiro.marvelhqs.presentation.interfaces
+
+interface Paginator<Key, Item> {
+    suspend fun loadNextItems()
+    fun reset()
+}
