+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presentation/ui/authentication/Credencials.kt
@@ -1,0 +1,7 @@
+package com.pinheiro.marvelhqs.presentation.ui.authentication
+
+data class Credencials(
+    var user: String = "",
+    var password: String = "",
+    var isVisble: Boolean = false,
+)
