+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presentation/ui/navigation/ComicNavigation.kt
@@ -1,0 +1,160 @@
+package com.pinheiro.marvelhqs.presentation.ui.navigation
+
+import android.annotation.SuppressLint
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.material.icons.filled.FavoriteBorder
+import androidx.compose.material.icons.filled.Home
+import androidx.compose.material.icons.filled.Output
+import androidx.compose.material.icons.outlined.Home
+import androidx.compose.material3.DrawerValue
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ModalDrawerSheet
+import androidx.compose.material3.ModalNavigationDrawer
+import androidx.compose.material3.NavigationDrawerItem
+import androidx.compose.material3.NavigationDrawerItemDefaults
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.rememberDrawerState
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.rememberNavController
+import com.pinheiro.marvelhqs.R
+import com.pinheiro.marvelhqs.presentation.ui.comic.ComicScreen
+import com.pinheiro.marvelhqs.presentation.ui.components.MarvelTopAppBar
+import com.pinheiro.marvelhqs.presentation.ui.favorite.FavoriteScreen
+import kotlinx.coroutines.launch
+
+@SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ComicNavigation(mainNavController: NavHostController) {
+    val items = listOf(
+        NavigationItem(
+            title = "All",
+            selectedIcon = Icons.Filled.Home,
+            unselectedIcon = Icons.Outlined.Home,
+            route = "home"
+
+        ),
+        NavigationItem(
+            title = stringResource(R.string.FAVORITE),
+            selectedIcon = Icons.Default.Favorite,
+            unselectedIcon = Icons.Default.FavoriteBorder,
+            route = "favorite"
+        ),
+        NavigationItem(
+            title = stringResource(R.string.LOG_OUT),
+            selectedIcon = Icons.Default.Output,
+            unselectedIcon = Icons.Default.Output,
+            route = "exit"
+        ),
+    )
+
+    val navController = rememberNavController()
+
+    Surface(
+        modifier = Modifier.fillMaxSize(),
+        color = MaterialTheme.colorScheme.background
+    ) {
+        val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
+        val scope = rememberCoroutineScope()
+        var selectedItemIndex by rememberSaveable {
+            mutableStateOf(0)
+        }
+
+        ModalNavigationDrawer(
+            drawerContent = {
+                ModalDrawerSheet {
+                    Spacer(modifier = Modifier.height(16.dp))
+                    items.forEachIndexed { index, item ->
+                        NavigationDrawerItem(
+                            label = {
+                                Text(text = item.title)
+                            },
+                            selected = index == selectedItemIndex,
+                            onClick = {
+                                navController.navigate(item.route)  {
+                                    popUpTo(0)
+                                }
+                                selectedItemIndex = index
+                                scope.launch {
+                                    drawerState.close()
+                                }
+                            },
+                            icon = {
+                                Icon(
+                                    imageVector = if (index == selectedItemIndex) {
+                                        item.selectedIcon
+                                    } else item.unselectedIcon,
+                                    contentDescription = item.title
+                                )
+                            },
+                            badge = {
+                                item.badgeCount?.let {
+                                    Text(text = item.badgeCount.toString())
+                                }
+                            },
+                            modifier = Modifier
+                                .padding(NavigationDrawerItemDefaults.ItemPadding)
+                        )
+
+                    }
+                }
+            },
+            drawerState = drawerState
+        ) {
+            Scaffold(
+                topBar = {
+                    MarvelTopAppBar(drawerState)
+        }
+
+            ) {
+
+                NavHost(navController = navController, startDestination = "home") {
+                    composable("home") {
+                        ComicScreen(mainNavController)
+                    }
+                    composable("favorite") {
+                        FavoriteScreen()
+                    }
+                    composable("exit") {
+//                        navController.popBackStack()
+                        mainNavController.navigate(Login) {
+                            popUpTo(0)
+                        }
+
+                    }
+                }
+
+            }
+        }
+    }
+}
+
+
+//data class NavigationItem(
+//    val title: String,
+//    val selectedIcon: ImageVector,
+//    val unselectedIcon: ImageVector,
+//    val badgeCount: Int? = null,
+//    val route: String,
+//
+//    )
