+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presentation/ui/navigation/NavigationItem.kt
@@ -1,0 +1,12 @@
+package com.pinheiro.marvelhqs.presentation.ui.navigation
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class NavigationItem(
+    val title: String,
+    val selectedIcon: ImageVector,
+    val unselectedIcon: ImageVector,
+    val badgeCount: Int? = null,
+    val route: String,
+
+    )
