+++ MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presentation/ui/viewmodels/AuthViewModel.kt
@@ -1,0 +1,6 @@
+package com.pinheiro.marvelhqs.presentation.ui.viewmodels
+
+import androidx.lifecycle.ViewModel
+
+class AuthViewModel: ViewModel() {
+}
