--- MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/interfaces/Paginator.kt
@@ -1,6 +1,0 @@
-package com.pinheiro.marvelhqs.presenter.interfaces
-
-interface Paginator<Key, Item> {
-    suspend fun loadNextItems()
-    fun reset()
-}
