--- MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/authentication/Credencials.kt
@@ -1,7 +1,0 @@
-package com.pinheiro.marvelhqs.presenter.ui.authentication
-
-data class Credencials(
-    var user: String = "",
-    var password: String = "",
-    var isVisble: Boolean = false,
-)
