--- MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/components/LogButton.kt
@@ -1,39 +1,0 @@
-package com.pinheiro.marvelhqs.presenter.ui.components
-
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
-import androidx.compose.material3.Button
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-
-@Composable
-fun LogButton(
-    text: String,
-    onClick: () -> Unit,
-) {
-    Column(  // 2
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(20.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-    ) {
-        Button(  // 3
-            onClick = { onClick() },
-            modifier = Modifier
-                .width(200.dp)
-                .height(50.dp),
-        ) {
-            Text(  // 4
-                text = text,
-                fontSize = 20.sp,
-            )
-        }
-    }
-}
