--- MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/components/MarvelTopAppBar.kt
@@ -1,35 +1,0 @@
-package com.pinheiro.marvelhqs.presenter.ui.components
-
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Menu
-import androidx.compose.material3.DrawerState
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.rememberCoroutineScope
-import kotlinx.coroutines.launch
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MarvelTopAppBar(drawerState: DrawerState) {
-    val scope = rememberCoroutineScope()
-   return TopAppBar(
-        title = {
-            Text(text = "Marvel")
-        },
-        navigationIcon = {
-            IconButton(onClick = {
-                scope.launch {
-                    drawerState.open()
-                }
-            }) {
-                Icon(
-                    imageVector = Icons.Default.Menu,
-                    contentDescription = "Menu"
-                )
-            }
-        })
-}
