--- MarvelHQs2/app/src/main/java/com/pinheiro/marvelhqs/presenter/ui/viewmodels/AuthViewModel.kt
@@ -1,6 +1,0 @@
-package com.pinheiro.marvelhqs.presenter.ui.viewmodels
-
-import androidx.lifecycle.ViewModel
-
-class AuthViewModel: ViewModel() {
-}
